refactor: グローバルナビ固定化 + App Router正規化 (Closes #62)
- plot-management.tsx の currentView 状態機械を廃止
- (main)/layout.tsx にグローバルサイドバー + ヘッダーを固定配置
- 全画面を独立ルートに移行 (/plots, /staff, /masters 等)
- PlotDetailSidebar の isPlotContextView 分岐を削除
- 区画詳細ページにパンくず + ツールバー(書類作成/履歴/削除)追加
- onViewChange prop を全コンポーネントから除去、useRouter に統一
- UserMenu を自律的にルーティング (router.push)
