+++ src/app/(main)/bulk-import/page.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { RoleGuard } from '@/components/auth-guard';
+import BulkImportPage from '@/components/bulk-import';
+
+export default function BulkImportRoute() {
+  return (
+    <RoleGuard requiredRoles={['manager', 'admin']}>
+      <BulkImportPage />
+    </RoleGuard>
+  );
+}
