+++ src/app/(main)/collective-burials/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import CollectiveBurialManagement from '@/components/collective-burial';
+
+export default function CollectiveBurialsPage() {
+  return <CollectiveBurialManagement />;
+}
