+++ src/app/(main)/documents/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { DocumentManagement } from '@/components/document-management';
+
+export default function DocumentsPage() {
+  return <DocumentManagement />;
+}
