+++ src/app/(main)/masters/page.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { RoleGuard } from '@/components/auth-guard';
+import MastersManagement from '@/components/masters-management';
+
+export default function MastersPage() {
+  return (
+    <RoleGuard requiredRoles={['admin']}>
+      <MastersManagement />
+    </RoleGuard>
+  );
+}
