+++ src/app/(main)/plot-availability/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import PlotAvailabilityManagement from '@/components/plot-availability-management';
+
+export default function PlotAvailabilityPage() {
+  return <PlotAvailabilityManagement />;
+}
