+++ src/app/(main)/plots/[id]/documents/page.tsx
@@ -1,0 +1,25 @@
+'use client';
+
+import { useParams, useRouter } from 'next/navigation';
+import { usePlotDetail } from '@/hooks/usePlots';
+import { DocumentManagement } from '@/components/document-management';
+
+export default function PlotDocumentHistoryPage() {
+  const params = useParams();
+  const router = useRouter();
+  const plotId = params.id as string;
+
+  const { plot } = usePlotDetail(plotId);
+
+  const plotCode = plot?.physicalPlot?.plotNumber || '';
+  const customerName = plot?.roles?.[0]?.customer?.name || plotCode;
+
+  return (
+    <DocumentManagement
+      customerId={plotId}
+      customerName={customerName}
+      initialMode="list"
+      onBack={() => router.push(`/plots/${plotId}`)}
+    />
+  );
+}
