+++ src/app/(main)/profile/page.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import ProfilePage from '@/components/profile-page';
+
+export default function ProfileRoute() {
+  const router = useRouter();
+
+  return (
+    <ProfilePage onBack={() => router.push('/plots')} />
+  );
+}
