+++ src/app/(main)/staff/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import StaffManagement from '@/components/staff-management';
+
+export default function StaffPage() {
+  return <StaffManagement />;
+}
