+++ src/components/collective-burial/Management.tsx
@@ -10,16 +10,13 @@ import CollectiveBurialListView from './ListView';
 import CollectiveBurialDetailView from './DetailView';
 import { useCollectiveBurialDetail } from '@/hooks/useCollectiveBurials';
 import { CollectiveBurialListItem } from '@/lib/api';
-import { ViewType } from '@/types/plot-detail';
-
 type ViewMode = 'list' | 'detail';
 
 interface CollectiveBurialManagementProps {
   onBack?: () => void;
-  onViewChange?: (view: ViewType) => void;
 }
 
-export default function CollectiveBurialManagement({ onBack, onViewChange }: CollectiveBurialManagementProps) {
+export default function CollectiveBurialManagement({ onBack }: CollectiveBurialManagementProps) {
   const [viewMode, setViewMode] = useState<ViewMode>('list');
   const [selectedId, setSelectedId] = useState<string | null>(null);
   const [listRefreshKey, setListRefreshKey] = useState(0);
@@ -51,7 +48,6 @@ export default function CollectiveBurialManagement({ onBack, onViewChange }: Col
         key={listRefreshKey}
         onBack={onBack}
         onSelectRecord={handleSelectRecord}
-        onViewChange={onViewChange}
       />
     );
   }
