+++ src/components/masters-management.tsx
@@ -17,7 +17,6 @@ import {
 } from '@/lib/api';
 import { showSuccess, showError } from '@/lib/toast';
 import PageHeader from '@/components/page-header';
-import { ViewType } from '@/types/plot-detail';
 
 function formatApiError(
   message: string,
@@ -48,11 +47,7 @@ const MASTER_TYPES: MasterTypeConfig[] = [
   { key: 'section-name', label: '区画名', dataKey: 'sectionName' },
 ];
 
-interface MastersManagementProps {
-  onViewChange?: (view: ViewType) => void;
-}
-
-export default function MastersManagement({ onViewChange }: MastersManagementProps) {
+export default function MastersManagement() {
   const { user } = useAuth();
   const isAdmin = user?.role === 'admin';
 
@@ -179,7 +174,6 @@ export default function MastersManagement({ onViewChange }: MastersManagementPro
             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.066 2.573c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.573 1.066c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.066-2.573c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37a1.724 1.724 0 002.573-1.066z" /><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
           </svg>
         }
-        onViewChange={onViewChange}
       />
 
       {isLoading ? (
