+++ src/components/plot-detail-sidebar/index.ts
@@ -1,2 +1,1 @@
-export { default as PlotDetailSidebar } from './PlotDetailSidebar';
 export * from './dialogs';
