+++ src/components/user-menu.tsx
@@ -10,7 +10,6 @@ import {
   DropdownMenuSeparator,
   DropdownMenuLabel,
 } from '@/components/ui/dropdown-menu';
-import { ViewType } from '@/types/plot-detail';
 
 const ROLE_LABELS: Record<string, string> = {
   admin: '管理者',
@@ -19,11 +18,7 @@ const ROLE_LABELS: Record<string, string> = {
   viewer: '閲覧者',
 };
 
-interface UserMenuProps {
-  onViewChange: (view: ViewType) => void;
-}
-
-export default function UserMenu({ onViewChange }: UserMenuProps) {
+export default function UserMenu() {
   const { user, logout, isLoading } = useAuth();
   const router = useRouter();
 
@@ -58,7 +53,7 @@ export default function UserMenu({ onViewChange }: UserMenuProps) {
           </div>
         </DropdownMenuLabel>
         <DropdownMenuSeparator />
-        <DropdownMenuItem onClick={() => onViewChange('profile')}>
+        <DropdownMenuItem onClick={() => router.push('/profile')}>
           <svg className="w-4 h-4 mr-2 text-hai" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
           </svg>
